Add Rails 3 support.
In addition to adding Rail 3 support, this commit also addresses an issue which
appeared and was preventing all tests from running. I am not sure what triggered
this issue to appear but including a require statement for the missing class in
the test helper cleared it up. Probably just a side effect of the hacky way I
have the tests setup.